tests/wpt/web-platform-tests/css/css-masking/parsing/mask-valid.sub.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-masking-1/#the-mask">
<meta name="assert" content="mask supports the full '<mask-layer>#' grammar.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
// <mask-layer> = <mask-reference> <masking-mode>? || <position> [ / <bg-size> ]? ||
// <repeat-style> || <geometry-box> || [ <geometry-box> | no-clip ] || <compositing-operator>

// <mask-reference> <masking-mode>?
// <mask-reference> = none | <image> | <mask-source>
// <masking-mode> = alpha | luminance | auto
test_valid_value('mask', 'none');
test_valid_value('mask', 'none alpha');
test_valid_value('mask', 'linear-gradient(to left bottom, red, blue)');
test_valid_value('mask', 'linear-gradient(to left bottom, red, blue) luminance');
test_valid_value('mask', 'url("https://{{host}}/")');
test_valid_value('mask', 'url("https://{{host}}/") alpha');

// <position> [ / <bg-size> ]?
test_valid_value('mask', 'linear-gradient(to left bottom, red, blue) 1px 2px');
test_valid_value('mask', 'url("https://{{host}}/") 1px 2px / contain');

// <repeat-style> = repeat-x | repeat-y | [repeat | space | round | no-repeat]{1,2}
test_valid_value('mask', 'none repeat-y');

// <geometry-box> = <shape-box> | fill-box | stroke-box | view-box
// <shape-box> = <box> | margin-box
// <box> = border-box | padding-box | content-box
test_valid_value('mask', 'none border-box', 'none');
test_valid_value('mask', 'linear-gradient(to left bottom, red, blue) padding-box');
test_valid_value('mask', 'none content-box');
test_valid_value('mask', 'none margin-box');
test_valid_value('mask', 'url("https://{{host}}/") fill-box');
test_valid_value('mask', 'linear-gradient(to left bottom, red, blue) stroke-box');
test_valid_value('mask', 'none view-box');

// [ <geometry-box> | no-clip ]
test_valid_value('mask', 'none no-clip', 'none border-box no-clip');

// <compositing-operator> = add | subtract | intersect | exclude
test_valid_value('mask', 'url("https://{{host}}/") add', 'url("https://{{host}}/")');
test_valid_value('mask', 'none subtract');
test_valid_value('mask', 'url("https://{{host}}/") intersect');
test_valid_value('mask', 'linear-gradient(to left bottom, red, blue) exclude');


// <mask-layer> = <mask-reference> <masking-mode>? || <position> [ / <bg-size> ]? ||
// <repeat-style> || <geometry-box> || [ <geometry-box> | no-clip ] || <compositing-operator>
test_valid_value('mask', 'intersect no-clip space round 1px 2px / contain stroke-box linear-gradient(to left bottom, red, blue) luminance', 'linear-gradient(to left bottom, red, blue) luminance 1px 2px / contain space round stroke-box no-clip intersect');

test_valid_value('mask', 'intersect no-clip space round 1px 2px / contain view-box, stroke-box linear-gradient(to left bottom, red, blue) luminance', 'none 1px 2px / contain space round view-box no-clip intersect, linear-gradient(to left bottom, red, blue) luminance stroke-box');
</script>
</body>
</html>
