LayoutTests/imported/w3c/web-platform-tests/css/css-masking/parsing/mask-valid.sub-expected.txt


PASS e.style['mask'] = "none" should set the property value
PASS e.style['mask'] = "linear-gradient(to left bottom, red, blue)" should set the property value
PASS e.style['mask'] = "linear-gradient(to left bottom, red, blue) luminance" should set the property value
PASS e.style['mask'] = "url(\"https://localhost/\")" should set the property value
PASS e.style['mask'] = "linear-gradient(to left bottom, red, blue) 1px 2px" should set the property value
PASS e.style['mask'] = "url(\"https://localhost/\") 1px 2px / contain" should set the property value
PASS e.style['mask'] = "repeat-y" should set the property value
FAIL e.style['mask'] = "border-box" should set the property value assert_equals: serialization should be canonical expected "none" but got "border-box"
PASS e.style['mask'] = "linear-gradient(to left bottom, red, blue) padding-box" should set the property value
PASS e.style['mask'] = "content-box" should set the property value
FAIL e.style['mask'] = "margin-box" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['mask'] = "url(\"https://localhost/\") fill-box" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['mask'] = "linear-gradient(to left bottom, red, blue) stroke-box" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['mask'] = "view-box" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['mask'] = "no-clip" should set the property value assert_equals: serialization should be canonical expected "no-clip" but got "border-box no-clip"
FAIL e.style['mask'] = "url(\"https://localhost/\") add" should set the property value assert_equals: serialization should be canonical expected "url(\"https://localhost/\")" but got "url(\"https://localhost/\") add"
PASS e.style['mask'] = "subtract" should set the property value
PASS e.style['mask'] = "url(\"https://localhost/\") intersect" should set the property value
PASS e.style['mask'] = "linear-gradient(to left bottom, red, blue) exclude" should set the property value
PASS e.style['mask'] = "alpha" should set the property value
PASS e.style['mask'] = "url(\"https://localhost/\") alpha" should set the property value
FAIL e.style['mask'] = "intersect no-clip space round 1px 2px / contain stroke-box linear-gradient(to left bottom, red, blue) luminance" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['mask'] = "intersect no-clip space round 1px 2px / contain view-box, stroke-box linear-gradient(to left bottom, red, blue) luminance" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['mask'] = "none alpha" should set the property value assert_equals: serialization should be canonical expected "alpha" but got "none alpha"
